"init command" don't overwrite config file.

+++ src/cmd/index.ts
@@ -50,19 +50,32 @@ export default {
    * @param name プロジェクト名
    */
   init(name: string) {
-    const paletteTemplateDir = path.join(
-      common.JSON_FILE_DIRECTORY,
-      common.CONFIG_FILE_NAME
-    );
-    const settingTemplate = fs.readFileSync(paletteTemplateDir, {
-      encoding: 'utf-8',
-    });
-    const settingJson = JSON.parse(settingTemplate);
-    settingJson.projectName = name;
-    fs.writeFileSync(
-      'colorpalette.config.json',
-      JSON.stringify(settingJson, null, 2)
-    );
+    if (fs.existsSync(path.join(common.root, common.CONFIG_FILE_NAME))) {
+      new Message('error', 'Project file has already been generated.');
+      new Message(
+        'log',
+        `config file here "${path.join(
+          common.root,
+          common.CONFIG_FILE_NAME
+        )}"\n`
+      );
+      return;
+    } else {
+      const paletteTemplateDir = path.join(
+        common.JSON_FILE_DIRECTORY,
+        common.CONFIG_FILE_NAME
+      );
+      const settingTemplate = fs.readFileSync(paletteTemplateDir, {
+        encoding: 'utf-8',
+      });
+      const settingJson = JSON.parse(settingTemplate);
+      settingJson.projectName = name;
+      fs.writeFileSync(
+        'colorpalette.config.json',
+        JSON.stringify(settingJson, null, 2)
+      );
+      return new Message('complete', `create new color-palette ${name}`);
+    }
   },
 
   /**
